+++ R/model.R
@@ -282,7 +282,7 @@ setMethod(
 #' @slot value_type type of values (real, integer, logical) (default real)
 #' @slot min_value minimal value (default NA)
 #' @slot max_value maximal value (default NA)
-#' @slot plot_color color in pot (default "")
+#' @slot plot_color color in plot (default "")
 #' @slot plot_line_width width of line in plot (default 1)
 #' @exportClass mc_Sensor
 #' @seealso [mc_data_sensors]
@@ -1016,7 +1016,7 @@ setMethod(
 
 .model_is_logger_type_hobo <- function(object, data, secondary_column, logger_type) {
     if(!is.na(object@logger_type)) {
-        return(FALSE)
+        return(object@logger_type == logger_type)
     }
     column_pattern <- .model_const_HOBO_LOGGER_TYPE_SECONDARY_TITLES[[logger_type]]
     parts <- stringr::str_match(data[[secondary_column]][[1]], column_pattern)
@@ -1042,7 +1042,7 @@ setMethod(
     if(length(object@columns) == 0) {
         return(FALSE)
     }
-    if(is.na(object@date_format)) {
+    if(all(is.na(object@date_format))) {
         warning(.model_const_MESSAGE_HOBO_DATE_FORMAT_PROBLEM)
         return(FALSE)
     }