+++ tests/testthat/test-read.R
@@ -85,6 +85,9 @@ test_that("mc_read_files HOBO", {
     data <- mc_read_files(c("../data/HOBO/20024354_comma.csv", "../data/HOBO/20024354_semicolon.txt", "../data/HOBO/20024354_tab.txt"),
                           "HOBO",  date_format = "%y.%m.%d %H:%M:%S", tz_offset = 120, clean=FALSE)
     test_raw_data_format(data)
+    data <- mc_read_files(c("../data/HOBO/20024354_comma.csv", "../data/HOBO/20024354_semicolon.txt", "../data/HOBO/20024354_tab.txt"),
+                          "HOBO",  date_format = "%y.%m.%d %H:%M:%S@%y.%m.%d", tz_offset = 120, clean=FALSE)
+    test_raw_data_format(data)
 })
 
 test_that("mc_read_data HOBO", {
@@ -275,9 +278,25 @@ test_that("mc_read_files user_data_formats auto datetime", {
 })
 
 test_that("mc_read_files TOMST custom date time format", {
-    expect_error(expect_warning(expect_warning(data <- mc_read_files("../data/TOMST-date", "TOMST"))))
-    data <- mc_read_files("../data/TOMST-date", "TOMST", date_format=c("%d.%m.%Y %H:%M:%S", "%d.%m.%Y"),
+    expect_error(expect_warning(expect_warning(data <- mc_read_files("../data/TOMST-date/data_91171153_2023_11_14_0.csv", "TOMST"))))
+    data <- mc_read_files("../data/TOMST-date/data_91171153_2023_11_14_0.csv", "TOMST", date_format=c("%d.%m.%Y %H:%M:%S", "%d.%m.%Y"),
                           silent=TRUE)
+    test_raw_data_format(data)
+    data <- mc_read_data("../data/TOMST-date/files_table.csv", clean=FALSE)
+    test_raw_data_format(data)
+    files_table <- read.table("../data/TOMST-date/files_table.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
+    data <- mc_read_data(files_table, clean=FALSE)
     test_raw_data_format(data)
 })
 
+test_that("mc_read_files HOBO logger_format", {
+    files_table <- as.data.frame(tibble::tribble(
+        ~path, ~locality_id, ~data_format, ~date_format, ~logger_type,
+             "../data/HOBO/20024354.txt", "A", "HOBO", "%d.%m.%Y %H:%M:%S", "HOBO_U23-001A",
+        "../data/HOBO/2015_10382557.txt", "A", "HOBO", "%d.%m.%Y %H:%M:%S", "HOBO_U23-004",
+    ))
+    data <- mc_read_data(files_table, clean=FALSE)
+    test_raw_data_format(data)
+    expect_equal(names(data$localities$A$loggers[[1]]$sensors), c("HOBO_T", "HOBO_RH"))
+    expect_equal(names(data$localities$A$loggers[[2]]$sensors), c("HOBO_T", "HOBO_extT"))
+})